Some casts.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@4768 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/ftp/ftp/security.c
@@ -436,7 +436,7 @@ void pbsz(int size)
 	buffer_size = size;
     }
     if(new != size)
-	reply(200, "PBSZ=%lu", new);
+	reply(200, "PBSZ=%lu", (unsigned long)new);
     else
 	reply(200, "OK");
 }
